FFS-2322: Consolidate Pinwheel helpers into PinwheelDataHelper (#422)
